[Serve] refactor pow 2 scheduler into common logics and mixins
